backend/requirements.txt
# ZEKO Backend API
# FastAPI + Firebase + Google Cloud AI

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Firebase Admin SDK
firebase-admin==6.2.0

# Google Cloud APIs
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
google-cloud-language==2.11.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2

# HTTP Client
httpx==0.25.2
aiofiles==23.2.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
pymongo==4.6.0
motor==3.3.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
